fix: refactor helpers and split isStatusMatch to Suite and Browser
