+++ lib/static/modules/group-errors.js
@@ -1,6 +1,7 @@
 import {get} from 'lodash';
 import {isTestNameMatchFilters} from './utils';
-import {shouldShowBrowser, getFailedSuiteResults} from './selectors/tree';
+import {shouldShowBrowser} from './selectors/helpers';
+import {getFailedSuiteResults} from './selectors/tree';
 import viewModes from '../../constants/view-modes';
 
 const imageComparisonErrorMessage = 'image comparison failed';
