+++ lib/static/modules/load-plugin.js
@@ -80,9 +80,13 @@ async function initPlugin(plugin, pluginName) {
             const depArgs = deps.map(dep => whitelistedDeps[dep]);
             // cyclic dep, resolve it dynamically
             const actions = await import('./actions');
-            const selectorsTree = await import('./selectors/tree');
-            const selectorsHelper = await import('./selectors/selectorsHelper');
-            return plugin(...depArgs, {pluginName, actions, selectors: {selectorsTree, selectorsHelper}});
+            const selectors = {
+                helpers: await import('./selectors/helpers'),
+                stats: await import('./selectors/stats'),
+                tree: await import('./selectors/tree'),
+                view: await import('./selectors/view')
+            };
+            return plugin(...depArgs, {pluginName, actions, selectors});
         }
 
         return plugin;
