+++ lib/static/modules/selectors/helpers.js
@@ -2,7 +2,7 @@ import {find, isEmpty} from 'lodash';
 import {isStatusMatchViewMode, getSuiteBrowsers} from './tree';
 
 export function shouldBrowserBeShown(browser, lastResultStatus, errorGroupBrowserIds, filteredBrowsers, viewMode) {
-    if (!isStatusMatchViewMode(lastResultStatus, viewMode)) {
+    if (!isBrowserMatchViewMode(browser, lastResultStatus, viewMode)) {
         return false;
     }
 
@@ -10,7 +10,7 @@ export function shouldBrowserBeShown(browser, lastResultStatus, errorGroupBrowse
 }
 
 export function shouldSuiteBeShown(suite, suites, browsers, errorGroupBrowserIds, testNameFilter, strictMatchFilter, filteredBrowsers, viewMode) {
-    if (!isStatusMatchViewMode(suite.status, viewMode)) {
+    if (!isSuiteMatchViewMode(suite, viewMode)) {
         return false;
     }
 
@@ -25,6 +25,14 @@ export function shouldSuiteBeShown(suite, suites, browsers, errorGroupBrowserIds
 
         return shouldShowBrowser(browser, filteredBrowsers, errorGroupBrowserIds);
     });
+}
+
+export function isBrowserMatchViewMode(browser, lastResultStatus, viewMode) {
+    return isStatusMatchViewMode(lastResultStatus, viewMode);
+}
+
+export function isSuiteMatchViewMode(suite, viewMode) {
+    return isStatusMatchViewMode(suite.status, viewMode);
 }
 
 export function shouldShowBrowser(browser, filteredBrowsers, errorGroupBrowserIds = []) {
