+++ lib/static/modules/selectors/tree.js
@@ -5,7 +5,7 @@ import viewModes from '../../../constants/view-modes';
 import {isFailStatus, isErroredStatus, isIdleStatus} from '../../../common-utils';
 import {isNodeFailed, isNodeSuccessful, isAcceptable} from '../utils';
 
-import {shouldSuiteBeShown, shouldBrowserBeShown} from './selectorsHelper';
+import {shouldSuiteBeShown, shouldBrowserBeShown} from './helpers';
 
 const getSuites = (state) => state.tree.suites.byId;
 const getBrowsers = (state) => state.tree.browsers.byId;
@@ -121,12 +121,14 @@ export const getFailedOpenedImageIds = createSelector(
     }
 );
 
+export const getVisibleRootSuiteIds = (rootSuiteIds, suites, ...args) => {
+    return rootSuiteIds.filter((suiteId) => shouldSuiteBeShown(suites[suiteId], suites, ...args));
+};
+
 export const mkGetVisibleRootSuiteIds = () => createSelector(
     getRootSuiteIds, getSuites, getBrowsers, getErrorGroupBrowserIds,
     getTestNameFilter, getStrictMatchFilter, getFilteredBrowsers, getViewMode,
-    (rootSuiteIds, suites, ...args) => {
-        return rootSuiteIds.filter((suiteId) => shouldSuiteBeShown(suites[suiteId], suites, ...args));
-    }
+    getVisibleRootSuiteIds
 );
 
 export const mkHasSuiteFailedRetries = () => createSelector(
